AI_Agent_Intermediate/requirements.txt
python-telegram-bot==21.0.1
openai==1.14.0
python-dotenv==1.0.1
supabase==2.3.6
langchain==0.1.13
langchain-community==0.0.29
langchain-huggingface==0.0.3
sentence-transformers==2.6.1
pymupdf==1.24.1
pydantic==2.6.4
